Restructure dtype-core & dtype-alias
