+++ packages/dtype-alias/index.js
@@ -1,4 +1,4 @@
 import AliasMeta from '../../contracts/build/contracts/Alias.json';
 
 export {AliasMeta};
-export * from './utils';
+export * from './src/utils';
