+++ packages/dtype-core/index.js
@@ -1,8 +1,8 @@
 import dTypeMeta from '../../contracts/build/contracts/dType.json';
 
 export {dTypeMeta};
-export * from './blockchain';
-export * from './constants';
-export * from './ethers-utils';
-export * from './dtype-utils';
-export * from './dtype';
+export * from './src/blockchain';
+export * from './src/constants';
+export * from './src/ethers-utils';
+export * from './src/dtype-utils';
+export * from './src/dtype';
