+++ packages/dtype-core/src/constants.js
