+++ packages/dtype-core/src/dtype-utils.js
