+++ packages/dtype-core/src/dtype.js
