+++ packages/dtype-core/src/metadata.js
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 export const StorageBase = {
   "contractName": "StorageBase",
   "abi": [
